Improve detection of Python to embed.
Set the PYTHON_EMBED to point to the prefix of the Python that one wants
to embed.  If this is not set, it uses the default `python` executable.
This still requires that the user set the LD_LIBRARY_PATH or
DYLD_LIBRARY_PATH appropriate env vars to load the correct Python
library.  Once found, the correct libraries and includes for the given
Python are used to embed the Python.